Revert "Merge the 64-bit changes from casacore"
This reverts commit ddd16a415b0868181733c230d7308b49936019b6.
